Fix the integration test now that messages are received reliably
It can no longer be an integration test because the inner `connection`
must be accessed to force a close.

blued usually (always?) sends 2 messages in response to our message. The
integration test asserted that a second message shouldn't be available
after dropping the XPC connection, but that isn't correct because the
stream is buffered and owned by the user. Dropping the XPC connection
only prevents new messages from being sent.